Uklanjanje usercontrole
