+++ Classroom schedule/Classroom schedule/Program.cs
@@ -40,6 +40,9 @@ namespace Classroom_schedule
             services.AddScoped<IClassroomBusiness, ClassroomBusiness>();
             services.AddScoped<Login>();
             services.AddScoped<Form1>();
+           
+            
+
         }
     }
 }
