+++ Classroom schedule/DataAccessLayer/Constant.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Shared.Models;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -8,6 +9,6 @@ namespace DataAccessLayer
 {
     public class Constant
     {
-        public static string connectionString = "Data Source=(localdb)\\ProjectsV13;Initial Catalog=ClassroomSchedule;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False";
+        public static string connectionString = "Data Source=(localdb)\\ProjectsV13;Initial Catalog=Classroom schedule;Integrated Security=True;Connect Timeout=30;Encrypt=False;TrustServerCertificate=False;ApplicationIntent=ReadWrite;MultiSubnetFailover=False";
     }
 }
